Remove util functions to EnumUtils
