+++ core/src/main/java/org/apache/calcite/adapter/enumerable/EnumerableCalc.java
@@ -136,7 +136,7 @@ public class EnumerableCalc extends Calc implements EnumerableRel {
                 Enumerator.class, inputJavaType),
             "inputEnumerator");
     Expression input =
-        RexToLixTranslator.convert(
+        EnumUtils.convert(
             Expressions.call(
                 inputEnumerator,
                 BuiltInMethod.ENUMERATOR_CURRENT.method),
